RestauranteCRUD/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
mysqlclient==2.2.4
pymysql==1.1.1
pydantic==2.7.1
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4